requirements.txt
altgraph==0.17
asgiref==3.2.8
astroid==2.4.2
autopep8==1.5.3
Django==3.0.7
django-taggit==1.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.6.0
PyInstaller==3.6
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
